Extend insert mode `Tab` keymap and start `<Leader>bD` keymap
The `<Tab>` and `<S-Tab>` keymaps work perfectly. Tab to jump to next
snippet sequence when editing a snippet. Nice and easy. Dubious API, but
I don't really care.

The `<Leader>bD` keymap is more dubious. Starting to work on resolving
\#3, but I'm not quite there yet---the new function doesn't close
`:help` buffers and seems to throw errors sometimes, so I'm leaving it
as an alternate binding for now.