feat(mason-lspconfig): add `deferred-install.lua`
Hell of a fucking long file. Not gonna describe it. It basically just
gives me deferred (rather than automatic) installation of
`mason-lspconfig` sources, though. Preferable IMO to having your system
loaded with all of the LSPs you may never want to use. It's not super
well optimized, and I might turn it into a plugin in the future. Works
for now though.

Writing that also necessitated a change to my pre-commit config and
resulted in a change to my `mason` config. Nothing huge though.
Switching from `stylua-github` to `stylua-system` is a pain, but not
impossible.