feat(plugins): add `nvim-autopairs`
- Added `windwp/nvim-autopairs`
- Removed `m4xshen/autoclose.nvim`

This is a really nice plugin. The configuration is a little daunting,
but mostly I just copied off their wiki, and life is golden. Worth it
though, because this solved a bunch of smaller issues I had with
`autoclose.nvim`: `<>` for generics and not less-than, not autoclosing
when there's a close pair in the same line, ...

I still need to add `||` pairs for Rust, but I'm lazy, so I'll leave
that for next week.

Closes #7.