fix!: 8r8king changes (primarily mason & lspconfig)
Due to
[breaking changes in mason & lspconfig](mason-org/mason-lspconfig.nvim#545),
my `deferred-install` script no longer works. I've decided to just scrap
it---ultimately, what did it save? a couple gigabytes and a couple error
notifications? It's not like I can't spare the storage or network traffic.

While I'm not happy to scrap my system, and maybe someday I'd like to share it
with other people in a way that works, right now it serves more as maintenance
costs than as a productive use of my time & effort. Glad to be free of it.

I assume that I am using nvim>=0.11. I can't be bothered to do backwards
compat for 0.10. Sucks to suck.